c2 31.08.2018

+++ src/main/java/geschool/persistence/interfaces/CoursDAO.java
@@ -20,6 +20,6 @@ public interface CoursDAO {
     List<Cours> findAll();
     Cours findByIdCours (Integer id);
     Cours rechercherCoursParLibelleClasse(String l);
-    Long verifCoursExist(String l);
+    int verifCoursExist(String l);
 //    Integer rechercherLeNombreClasseCours(Integer id);
 }
